requirements.txt
alembic==1.3.3
amqp==2.5.2
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
bcrypt==3.1.7
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==5.0.3
fakeredis==1.1.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Migrate==1.8.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.8
importlib-metadata==1.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.7
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mockredis==0.1.3.dev0
more-itertools==8.1.0
nose==1.3.7
nose2==0.9.1
packaging==20.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.4
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
pytz==2019.3
redis==3.3.11
requests==2.22.0
requests-cache==0.5.2
six==1.14.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
zipp==1.0.0
